requirements.txt
Django~=5.0.7
pillow~=10.3.0
typing_extensions~=4.12.2
ipython~=8.24.0
pip~=24.0
attrs~=23.2.0
distro~=1.9.0
wheel~=0.43.0
docutils~=0.21.2
Sphinx~=7.3.7
asyncio~=3.4.3
tornado~=6.4
Jinja2~=3.1.4
asgiref~=3.8.1
sqlparse~=0.5.0
psycopg2~=2.9.9
colorama~=0.4.6
requests~=2.32.3